refactor(seances): SeanceCard sous 300 lignes — éclatement complet (H6)
Éclatement de SeanceCard.vue (567 l) en orchestrateur + sous-composants
présentationnels, CSS déplacé VERBATIM :
- SeanceCard.vue 567→53 (orchestrateur, CSS .seance-card)
- SeanceCardHeader.vue (137) titre + badge statut visio (+@keyframes pulse)
- SeanceCardInfo.vue (103) grille date/horaire/classe/salle
- SeanceCardActions.vue (288) 4 états visio
- _seance-card-buttons.scss (68) chrome boutons partagé, @use'd scoped
Tests: SeanceCard.test.js (parité, intact) + 11 verts. vite build vert.
Dette pré-existante conservée: classes utilitaires Tailwind inertes
(text-blue-700, flex…) sans pipeline Tailwind — laissées à l'identique.
